feat: 🎸 Code review in progress
Code review. Move retrieveMEtadata load to separate folders. Some minor
code clean up
